+++ src/main/java/lotto/application/RetrieveLottoUseCase.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.application;
 
 import java.util.List;
+import lotto.domain.Lotto;
 
 public interface RetrieveLottoUseCase {
