+++ src/main/java/lotto/common/RandomNumbersCreator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.common;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
