+++ src/main/java/lotto/domain/repository/InMemoryLottoRepository.java
@@ -1,7 +1,8 @@
-package lotto;
+package lotto.domain.repository;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.domain.Lotto;
 
 public class InMemoryLottoRepository implements LottoRepository {
