+++ src/main/java/lotto/domain/repository/LottoRepository.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.domain.repository;
 
 import java.util.List;
+import lotto.domain.Lotto;
 
 public interface LottoRepository {
