+++ src/main/java/lotto/view/converter/MessageConverter.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view.converter;
 
 import java.util.List;
